Services/DB_Handler/requirements.txt
sqlalchemy
alembic
psycopg2-binary
